+++ src/app/client.hpp
@@ -3,25 +3,18 @@
 
 #include "event_loop.hpp"
 #include "buttons.hpp"
-// IWYU pragma: no_forward_declare app::buttons
 #include "pen.hpp"
-// IWYU pragma: no_forward_declare app::pen
 #include "screen.hpp"
 #include "touch.hpp"
-// IWYU pragma: no_forward_declare app::touch
 #include <iosfwd>
 #include <optional>
 #include <poll.h> // IWYU pragma: keep
 #include <rfb/rfbclient.h>
-// IWYU pragma: no_include "rfb/rfbproto.h"
 #include <vector>
 
 namespace rmioc
 {
-    class screen;
-    class buttons;
-    class pen;
-    class touch;
+    class device;
 }
 
 namespace app
@@ -38,26 +31,20 @@ public:
      *
      * @param ip IP address of the VNC server to connect to.
      * @param port Port of the VNC server to connect to.
-     * @param screen_device Screen to update with data from the VNC server.
-     * @param buttons_device Buttons device to read input events from, or
-     * nullptr to ignore input events from the buttons.
-     * @param pen_device Pen digitizer to read input events from, or
-     * nullptr to ignore input events from the pen.
-     * @param touch_device Touchscreen device to read input events from, or
-     * nullptr to ignore input events from the touchscreen.
+     * @param device Handle to opened devices.
      */
-    client(
-        const char* ip, int port,
-        rmioc::screen& screen_device,
-        rmioc::buttons* buttons_device,
-        rmioc::pen* pen_device,
-        rmioc::touch* touch_device
-    );
+    client(const char* ip, int port, rmioc::device& device);
 
+    /** Disconnect the VNC client. */
     ~client();
 
-    /** Start the client event loop.  */
-    void event_loop();
+    /**
+     * Start the client event loop.
+     *
+     * @return True if the loop was exited because of a user action, false if
+     * it was because the server closed the connection.
+     */
+    bool event_loop();
 
 private:
     /** List of file descriptors to watch in the event loop. */
@@ -79,7 +66,7 @@ private:
     rfbClient* vnc_client;
 
     /** Event handler for the screen device. */
-    screen screen_handler;
+    std::optional<screen> screen_handler;
 
     /** Event handler for the buttons device. */
     std::optional<buttons> buttons_handler;