+++ src/app/screen.cpp
@@ -2,27 +2,39 @@
 #include "event_loop.hpp"
 #include "../log.hpp"
 #include "../rmioc/screen.hpp"
-#include "../rmioc/mxcfb.hpp"
 #include <algorithm>
 #include <chrono>
+#include <climits>
+#include <cstring>
 #include <ostream>
 #include <stdexcept>
 #include <rfb/rfbclient.h>
 // IWYU pragma: no_include <type_traits>
-// IWYU pragma: no_include "rfb/rfbproto.h"
 
 namespace chrono = std::chrono;
 
 /**
- * Time to wait between two repaints.
+ * Time to wait between two standard repaints.
  *
  * VNC servers tend to send a lot of small updates in a short period of time.
  * This delay allows grouping those small updates into a larger screen update.
  */
-constexpr chrono::milliseconds repaint_delay{400};
+constexpr chrono::milliseconds standard_repaint_delay{400};
+
+/**
+ * Time to wait between two fast repaints.
+ *
+ * This mode is used when the pen is active so that the user can have a quicker
+ * feedback on what they are drawing. When the pen is lifted, a high fidelity
+ * update is performed.
+ */
+constexpr chrono::milliseconds fast_repaint_delay{50};
 
 namespace app
 {
+
+// NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast,cppcoreguidelines-avoid-non-const-global-variables,cppcoreguidelines-avoid-magic-numbers)
+void* screen::instance_tag = reinterpret_cast<void*>(6803);
 
 screen::screen(rmioc::screen& device, rfbClient* vnc_client)
 : device(device)
@@ -31,14 +43,25 @@ screen::screen(rmioc::screen& device, rfbClient* vnc_client)
 {
     rfbClientSetClientData(
         this->vnc_client,
-        // ↓ Use of C library
-        // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-        reinterpret_cast<void*>(screen::instance_tag),
+        screen::instance_tag,
         this
     );
 
+    // Ask the server to send pixels in the same format as the screen buffer
+    this->vnc_client->format.bitsPerPixel = this->device.get_bits_per_pixel();
+    this->vnc_client->format.depth = this->device.get_bits_per_pixel();
+    this->vnc_client->format.redShift = this->device.get_red_format().offset;
+    this->vnc_client->format.redMax = this->device.get_red_format().max();
+    this->vnc_client->format.greenShift = this->device.get_green_format().offset;
+    this->vnc_client->format.greenMax = this->device.get_green_format().max();
+    this->vnc_client->format.blueShift = this->device.get_blue_format().offset;
+    this->vnc_client->format.blueMax = this->device.get_blue_format().max();
+
+    // Force the raw encoding and override the rect reception methods
+    this->vnc_client->appData.encodingsString = "raw";
     this->vnc_client->MallocFrameBuffer = screen::create_framebuf;
-    this->vnc_client->GotFrameBufferUpdate = screen::recv_framebuf;
+    this->vnc_client->GotBitmap = screen::recv_update;
+    this->vnc_client->GotFrameBufferUpdate = screen::commit_updates;
 }
 
 void screen::repaint()
@@ -50,7 +73,7 @@ void screen::repaint()
         this->update_info.has_update = false;
     }
 
-    this->last_repaint_time = chrono::steady_clock::now();
+    this->last_repaint = chrono::steady_clock::now();
 
     log::print("Screen update")
         << this->update_info.w << 'x' << this->update_info.h << '+'
@@ -60,8 +83,8 @@ void screen::repaint()
         this->update_info.x, this->update_info.y,
         this->update_info.w, this->update_info.h,
         this->repaint_mode == repaint_modes::standard
-            ? mxcfb::waveform_modes::gc16
-            : mxcfb::waveform_modes::du
+            ? rmioc::waveform_modes::gl16
+            : rmioc::waveform_modes::du
     );
 }
 
@@ -88,30 +111,28 @@ auto screen::event_loop() -> event_loop_status
 {
     if (this->update_info.has_update)
     {
-        if (this->repaint_mode == repaint_modes::fast)
+        auto next_update_time = this->last_repaint + (
+            this->repaint_mode == repaint_modes::standard
+            ? standard_repaint_delay
+            : fast_repaint_delay
+        );
+
+        auto now = chrono::steady_clock::now();
+        long wait_time = chrono::duration_cast<chrono::milliseconds>(
+            next_update_time - now
+        ).count();
+
+        if (wait_time <= 0)
         {
             this->repaint();
         }
         else
         {
-            auto now = chrono::steady_clock::now();
-            int remaining_wait_time =
-                chrono::duration_cast<chrono::milliseconds>(
-                    this->last_repaint_time + repaint_delay - now
-                ).count();
-
-            if (remaining_wait_time <= 0)
-            {
-                this->repaint();
-            }
-            else
-            {
-                // Wait until the next update is due
-                return {
-                    /* quit = */ false,
-                    /* timeout = */ remaining_wait_time
-                };
-            }
+            // Wait until the next update is due
+            return {
+                /* quit = */ false,
+                /* timeout = */ wait_time
+            };
         }
     }
 
@@ -124,49 +145,87 @@ auto screen::create_framebuf(rfbClient* vnc_client) -> rfbBool
     auto* that = reinterpret_cast<screen*>(
         rfbClientGetClientData(
             vnc_client,
-            // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-            reinterpret_cast<void*>(screen::instance_tag)
+            screen::instance_tag
         ));
 
-    // Make sure the server gives us a compatible format
-    int xres_mem = static_cast<int>(that->device.get_xres_memory());
-    int yres_mem = static_cast<int>(that->device.get_yres_memory());
+    int xres = static_cast<int>(that->device.get_xres());
+    int yres = static_cast<int>(that->device.get_yres());
 
-    if (vnc_client->width < 0
-        || vnc_client->height < 0
-        || vnc_client->width != xres_mem
-        || vnc_client->height > yres_mem)
+    if (vnc_client->width < 0 || vnc_client->height < 0)
     {
         std::stringstream msg;
-        msg << "Server uses an unsupported resolution ("
-            << that->vnc_client->width << 'x' << that->vnc_client->height
-            << "). This client can only cope with a screen width of exactly "
-            << xres_mem << " pixels and a screen height of "
-            << yres_mem << " pixels";
+        msg << "Invalid server resolution ("
+            << vnc_client->width << 'x' << vnc_client->height;
         throw std::runtime_error{msg.str()};
     }
 
-    // Configure connection with device framebuffer settings
-    that->vnc_client->frameBuffer = that->device.get_data();
-    that->vnc_client->format.bitsPerPixel = that->device.get_bits_per_pixel();
-    that->vnc_client->format.redShift = that->device.get_red_offset();
-    that->vnc_client->format.redMax = that->device.get_red_max();
-    that->vnc_client->format.greenShift = that->device.get_green_offset();
-    that->vnc_client->format.greenMax = that->device.get_green_max();
-    that->vnc_client->format.blueShift = that->device.get_blue_offset();
-    that->vnc_client->format.blueMax = that->device.get_blue_max();
+    if (vnc_client->width > xres || vnc_client->height > yres)
+    {
+        std::cerr << "Warning: The server resolution ("
+            << vnc_client->width << 'x' << vnc_client->height
+            << ") does not fit in the screen ("
+            << xres << 'x' << yres << ")\nThe image will be cropped to fit\n";
+    }
 
     return TRUE;
 }
 
-void screen::recv_framebuf(rfbClient* vnc_client, int x, int y, int w, int h)
+void screen::recv_update(
+    rfbClient* vnc_client, const uint8_t* buffer,
+    int x, int y, int w, int h
+)
 {
     // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
     auto* that = reinterpret_cast<screen*>(
         rfbClientGetClientData(
             vnc_client,
-            // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
-            reinterpret_cast<void*>(screen::instance_tag)
+            screen::instance_tag
+        ));
+
+    if (x < 0 || y < 0
+        || x >= that->device.get_xres_memory()
+        || y >= that->device.get_yres_memory())
+    {
+        return;
+    }
+
+    std::size_t pixel_size = that->device.get_bits_per_pixel() / CHAR_BIT;
+
+    std::size_t dest_stride = that->device.get_xres_memory() * pixel_size;
+    std::size_t dest_size = dest_stride * that->device.get_yres_memory();
+    uint8_t* const dest = that->device.get_data();
+
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+    uint8_t* dest_line = dest + y * dest_stride + x * pixel_size;
+
+    std::size_t buffer_stride = w * pixel_size;
+    std::size_t buffer_size = buffer_stride * h;
+    const uint8_t* buffer_line = buffer;
+
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+    while (dest_line < dest + dest_size && buffer_line < buffer + buffer_size)
+    {
+        std::memcpy(
+            dest_line,
+            buffer_line,
+            std::min(dest_stride - x * pixel_size, buffer_stride)
+        );
+
+        // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+        dest_line += dest_stride;
+
+        // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-pointer-arithmetic)
+        buffer_line += buffer_stride;
+    }
+}
+
+void screen::commit_updates(rfbClient* vnc_client, int x, int y, int w, int h)
+{
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    auto* that = reinterpret_cast<screen*>(
+        rfbClientGetClientData(
+            vnc_client,
+            screen::instance_tag
         ));
 
     // Register the region as pending update, potentially extending