+++ src/rmioc/buttons.cpp
@@ -1,14 +1,29 @@
 #include "buttons.hpp"
 #include <vector>
+#include <fcntl.h>
 #include <linux/input-event-codes.h>
 #include <linux/input.h>
 
 namespace rmioc
 {
 
-buttons::buttons()
-: input("/dev/input/event2")
+buttons::buttons(const char* device_path)
+: input(device_path)
 {}
+
+auto buttons::is(const char* device_path) -> bool
+{
+    file_descriptor input_fd{device_path, O_RDONLY};
+    auto supp_events = supported_input_events(input_fd);
+
+    if (!supp_events.has_key())
+    {
+        return false;
+    }
+
+    auto supp_keys = supported_key_types(input_fd);
+    return supp_keys.has_power();
+}
 
 auto buttons::process_events() -> bool
 {