+++ src/rmioc/input.hpp
@@ -1,6 +1,9 @@
 #ifndef RMIOC_INPUT_HPP
 #define RMIOC_INPUT_HPP
 
+#include "flags.hpp"
+#include "file.hpp"
+#include <utility>
 #include <vector>
 #include <linux/input.h>
 
@@ -8,6 +11,35 @@ struct pollfd;
 
 namespace rmioc
 {
+
+/** Available Linux input events. */
+RMIOC_FLAGS_DEFINE(
+    input_events,
+    syn, key, rel, abs
+);
+
+/** Get the set of input events that can be emitted by a device. */
+input_events supported_input_events(int input_fd);
+
+/** Available Linux key types. */
+RMIOC_FLAGS_DEFINE(
+    key_types,
+    power, tool_pen, tool_rubber
+);
+
+/** Get the set of key codes that can be emitted by a device. */
+key_types supported_key_types(int input_fd);
+
+/** Available Linux absolute axes. */
+RMIOC_FLAGS_DEFINE(
+    abs_types,
+    x, y, pressure, distance, tilt_x, tilt_y,
+    mt_slot, mt_tracking_id, mt_position_x, mt_position_y,
+    mt_pressure, mt_orientation
+);
+
+/** Get the set of absolute axes that are supported by a device. */
+abs_types supported_abs_types(int input_fd);
 
 /**
  * Generic class for reading Linux input devices.
@@ -24,9 +56,6 @@ public:
      * @param device_path Path to the device.
      */
     input(const char* device_path);
-
-    /** Close the input device. */
-    ~input();
 
     /**
      * Setup a pollfd structure to wait for events on the device.
@@ -47,9 +76,17 @@ protected:
      */
     std::vector<input_event> fetch_events();
 
+    /**
+     * Get the minimum and maximum value of an absolute axis of the device.
+     *
+     * @param type Type of axis to query (see <linux/input.h> for a list)
+     * @return Axis limits.
+     */
+    std::pair<int, int> get_axis_limits(unsigned int type) const;
+
 private:
     /** File descriptor for the input device. */
-    int input_fd;
+    file_descriptor input_fd;
 
     /** List of queued events. */
     std::vector<input_event> queued_events;