+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v0.4.1
+
+- Fix support for release 2.9.1 on reMarkable 1.
+
+## v0.4.0
+
+- Add reMarkable 2 support (depends on [remarkable2-framebuffer](https://github.com/ddvk/remarkable2-framebuffer)).
+- Tolerate server resolutions differing from 1404x1872 or 1408x1872.
+    - The server image is cropped to fit in the screen.
+- Add `vnsee-gui` script to start VNSee from a GUI.
+    - This is a [simple](https://rmkit.dev/apps/sas) script.
+
+## v0.3.1
+
+- Fix compatibility with TigerVNC.
+    - Fill-in the `depth` PixelFormat flag appropriately.
+
 ## v0.3.0
 
 - Rename from rmvncclient to VNSee.